// payload-types.ts
export interface Event {
  id: number;
  title: string;
  game: 'ea-fc-27' | 'gta-vi' | 'battlefield-6' | 'call-of-duty';
  /**
   * Leer lassen, wenn der Termin noch nicht feststeht.
   */
  date?: string | null;
  description: string;
  updatedAt: string;
  createdAt: string;
}